Project generator

+++ src/Player.java
@@ -3,7 +3,7 @@ import java.util.ArrayList;
 public class Player {
     public String name;
     public String[] skillList = {"DATABASE","FRONTEND","WORDPRESS","PRESTASHOP"};
-    public Double cash;
+    public Integer cash;
     public ArrayList<Programmer> programmerList = new ArrayList<>();
     public Integer sellers;
     public Integer testers;
